+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 6.6.60
-
 - Core: let cron jobs run in default time zone
 - Core: only show yes no field in detail panel when value is set
+- Core: Fixed deadlock problem with JMAP sync modseq values
 
 25-03-2022 6.6.59
 - Core: bugfix saving ACL groups